Create Order API added

+++ src/main/java/com/example/estore/dto/ProductDTO.java
@@ -7,7 +7,7 @@ public class ProductDTO {
     private Long id;
     private String name;
     private String description;
-    private double price; // In rupees
+    private int price; // In paise
     private int stockQuantity;
     private Long sellerId; // The ID of the seller
     private String image; // Base64 encoded image data
